requirements.txt
beautifulsoup4==4.12.2
python-docx==1.1.0
lxml==4.9.3
requests==2.31.0
